+++ src/main/java/priv/hsy/redenvelops/service/UserService.java
@@ -3,17 +3,18 @@ package priv.hsy.redenvelops.service;
 import com.baomidou.mybatisplus.core.conditions.Wrapper;
 import priv.hsy.redenvelops.entity.User;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface UserService {
 
     List<User> selectAll();
 
-    User selectById(Integer id);
+    User selectById(BigInteger id);
 
     boolean updateById(User user);
 
     int delete(Wrapper<User> wrapper);
 
-    String updateUserinfo(int uid, double money);
+    String updateUserinfo(BigInteger uid, double money);
 }
